+++ apps/docs/versions/0.3.0/_astro/dialog.astro_astro_type_script_index_0_lang.D_bgSerk.js
@@ -1,0 +1,1 @@
+import{i}from"./dialog.CxuwDFnT.js";i();