+++ apps/docs/versions/0.3.0/_astro/first-screen.astro_astro_type_script_index_0_lang.BBmE5H-A.js
@@ -1,0 +1,1 @@
+import{i as a}from"./data-table.nmCFm2fH.js";a();