+++ apps/docs/versions/0.3.0/_astro/inline-editing.astro_astro_type_script_index_0_lang.D0R0OlO7.js
@@ -1,0 +1,1 @@
+import{i}from"./row-edit.CApkEEOf.js";i();