+++ apps/docs/versions/0.3.0/_astro/list-report.astro_astro_type_script_index_0_lang.jx-yCMVA.js
@@ -1,0 +1,1 @@
+import{i as t}from"./data-table.nmCFm2fH.js";t();document.getElementById("il-bulk")?.addEventListener("submit",e=>e.preventDefault());