+++ apps/docs/versions/0.3.0/_astro/login.astro_astro_type_script_index_0_lang.CPwoPeF7.js
@@ -1,0 +1,1 @@
+import{i}from"./validation-summary.aj4wCQdG.js";i();