+++ apps/docs/versions/0.3.0/_astro/money.astro_astro_type_script_index_0_lang.BBj79BDe.js
@@ -1,0 +1,1 @@
+import{i}from"./money-field.pfQeO_JL.js";import"./decimal-input.CirVt_Ss.js";i();