+++ apps/docs/versions/0.3.0/_astro/quantity.astro_astro_type_script_index_0_lang.80sSsT2D.js
@@ -1,0 +1,1 @@
+import{i}from"./quantity.O_2XHnPf.js";import"./decimal-input.CirVt_Ss.js";i();