+++ apps/docs/versions/0.3.0/_astro/table-toolbar.astro_astro_type_script_index_0_lang.CD-zbSrx.js
@@ -1,0 +1,1 @@
+import{i as a}from"./data-table.nmCFm2fH.js";import{a as i,i as t}from"./table-toolbar.BC_07usu.js";a();i();t();